Add session interface and native implementation
Web applications will often need some kind of a session to maintain
state between various stages of an operation or to store auth
credentials. In order to keep this code testable we must avoid the use
of the global session array. By providing both an interface and a
native implementation both goals can be achieved.

Includes a configuration that uses the native session as the default.